Refactor Unit tests

+++ tests/Unit/inc/Engine/Cache/WPCache/maybePreventDeactivation.php
@@ -4,7 +4,6 @@ namespace WP_Rocket\Tests\Unit\inc\Engine\Cache\WPCache;
 use WP_Rocket\Engine\Cache\WPCache;
 use WP_Rocket\Tests\Unit\FilesystemTestCase;
 use Brain\Monkey\Filters;
-use Brain\Monkey\Functions;
 
 /**
  * @covers \WP_Rocket\Engine\Cache\WPCache::maybe_prevent_deactivation
@@ -15,28 +14,20 @@ use Brain\Monkey\Functions;
 class Test_MaybePreventDeactivation extends FilesystemTestCase {
 	protected $path_to_test_data = '/inc/Engine/Cache/WPCache/maybePreventDeactivation.php';
 
-	public function testShouldBailOutWhenConfigFileFound() {
+	/**
+	 * @dataProvider providerTestData
+	 */
+	public function testShouldMaybePreventDeactivation( $config, $expected ) {
 		$wp_cache = new WPCache( $this->filesystem );
 
-		$this->assertSame( [], $wp_cache->maybe_prevent_deactivation( [] ) );
+		if ( ! $config['file_exist'] ) {
+			$this->filesystem->delete( 'vfs://public/wp-config.php' );
+		}
+
+		if ( isset( $config['set_filter_to_false'] ) && $config['set_filter_to_false'] ) {
+			Filters\expectApplied( 'rocket_set_wp_cache_constant' )->andReturn( false );
+		}
+
+		$this->assertSame( $expected, $wp_cache->maybe_prevent_deactivation( [] ) );
 	}
-
-	public function testShouldRBailOutWhenSetCacheConstFilterTrue() {
-		$this->filesystem->delete('vfs://public/wp-config.php' );
-		$wp_cache = new WPCache( $this->filesystem );
-
-		Filters\expectApplied( 'rocket_set_wp_cache_constant' )->andReturn( false );
-
-		$this->assertSame( [], $wp_cache->maybe_prevent_deactivation( [] ) );
-	}
-
-	public function testShouldAddCauseToCausesWhenNotPrevented() {
-		$this->filesystem->delete('vfs://public/wp-config.php' );
-		$wp_cache = new WPCache( $this->filesystem );
-
-		Filters\expectApplied( 'rocket_set_wp_cache_constant' )->andReturn( true );
-
-		$this->assertSame( [ 'wpconfig' ], $wp_cache->maybe_prevent_deactivation( [] ) );
-	}
-
 }
